Split Boost tests into directories
