+++ projects/plugins/boost/tests/e2e/specs/base/get-started.test.ts
@@ -1,7 +1,7 @@
 import { test, expect } from 'jetpack-e2e-commons/fixtures/base-test.js';
 import playwrightConfig from 'jetpack-e2e-commons/playwright.config.cjs';
-import { boostPrerequisitesBuilder } from '../lib/env/prerequisites.js';
-import { JetpackBoostPage } from '../lib/pages/index.js';
+import { boostPrerequisitesBuilder } from '../../lib/env/prerequisites.js';
+import { JetpackBoostPage } from '../../lib/pages/index.js';
 
 let jetpackBoostPage;
