+++ projects/plugins/boost/tests/e2e/specs/critical-css/critical-css.test.js
@@ -1,6 +1,6 @@
 import { test, expect } from 'jetpack-e2e-commons/fixtures/base-test.js';
-import { boostPrerequisitesBuilder } from '../lib/env/prerequisites.js';
-import { JetpackBoostPage } from '../lib/pages/index.js';
+import { boostPrerequisitesBuilder } from '../../lib/env/prerequisites.js';
+import { JetpackBoostPage } from '../../lib/pages/index.js';
 import { PostFrontendPage } from 'jetpack-e2e-commons/pages/index.js';
 import { DashboardPage, ThemesPage, Sidebar } from 'jetpack-e2e-commons/pages/wp-admin/index.js';
 import playwrightConfig from 'jetpack-e2e-commons/playwright.config.cjs';
