+++ projects/plugins/boost/tests/e2e/specs/modules/modules-common.test.js
@@ -1,6 +1,6 @@
-import { boostPrerequisitesBuilder } from '../lib/env/prerequisites.js';
+import { boostPrerequisitesBuilder } from '../../lib/env/prerequisites.js';
 import { test, expect } from 'jetpack-e2e-commons/fixtures/base-test.js';
-import { JetpackBoostPage } from '../lib/pages/index.js';
+import { JetpackBoostPage } from '../../lib/pages/index.js';
 import playwrightConfig from 'jetpack-e2e-commons/playwright.config.cjs';
 
 const modules = [
