+++ projects/plugins/boost/tests/e2e/specs/modules/score-auto-refresh.test.js
@@ -1,7 +1,7 @@
 import { test, expect } from 'jetpack-e2e-commons/fixtures/base-test.js';
-import { boostPrerequisitesBuilder } from '../lib/env/prerequisites.js';
+import { boostPrerequisitesBuilder } from '../../lib/env/prerequisites.js';
 import playwrightConfig from 'jetpack-e2e-commons/playwright.config.cjs';
-import { JetpackBoostPage } from '../lib/pages/index.js';
+import { JetpackBoostPage } from '../../lib/pages/index.js';
 
 test.describe( 'Auto refresh of speed scores', () => {
 	let page;
